refactor(WorldParticles)
- Move AnimBy enum into YMotion mode
- Convert YMotion object to a class
