+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/render/customambience/worldeffects/WorldParticlesMode.kt
@@ -22,6 +22,7 @@ package net.ccbluex.liquidbounce.features.module.modules.render.customambience.w
 import net.ccbluex.liquidbounce.config.types.group.Mode
 import net.ccbluex.liquidbounce.config.types.group.ModeValueGroup
 import net.ccbluex.liquidbounce.config.types.group.ToggleableValueGroup
+import net.ccbluex.liquidbounce.config.types.list.Tagged
 import net.ccbluex.liquidbounce.event.events.WorldRenderEvent
 import net.ccbluex.liquidbounce.event.handler
 import net.ccbluex.liquidbounce.features.module.modules.render.customambience.worldeffects.WorldParticles.coords
@@ -46,9 +47,9 @@ abstract class WorldParticlesMode(name: String) : Mode(name) {
     protected val animCurve by easing("AnimCurve", Easing.QUAD_IN_OUT)
     protected val canBeCovered by boolean("CanBeCovered", true)
 
-    protected object Ymotion : ToggleableValueGroup(WorldParticlesSimple, "YMotion", false) {
+    protected class Ymotion : ToggleableValueGroup(WorldParticlesSimple, "YMotion", false) {
         val motion by float("Motion", 2f, -10f..10f)
-        val animBy by enumChoice("AnimBy", AnimBy.AGE)
+        val animBy by enumChoice("AnimBy", AnimBy.PROGRESS)
     }
 
     protected var lastSpawnTime = 0L
@@ -77,12 +78,12 @@ abstract class WorldParticlesMode(name: String) : Mode(name) {
                 withPositionRelativeToCamera(it.value) {
                     poseStack.withPush {
 
-                        if (Ymotion.enabled) {
-                            val anim = when (Ymotion.animBy) {
+                        if (Ymotion().enabled) {
+                            val anim = when (Ymotion().animBy) {
                                 AnimBy.PROGRESS -> progress
                                 AnimBy.AGE -> age
                             }
-                            translate(0F, Ymotion.motion * anim, 0F)
+                            translate(0F, Ymotion().motion * anim, 0F)
                         }
 
                         mulPose(rot)
@@ -92,4 +93,13 @@ abstract class WorldParticlesMode(name: String) : Mode(name) {
             }
         }
     }
+    
+    @Suppress("unused")
+    enum class AnimBy(
+        override val tag: String,
+    ) : Tagged {
+        PROGRESS("Progress"),
+        AGE("Age");
+    }
+
 }
